Type integer temporaries, and implement the bitwise operators
Bitwise had been refused for want of type information: x & y on a temp the
decompiler believes is a float does not compile. The information already
existed. IntegerOperandAnalysis keeps a set of register components that only
ever hold integers, so TempVariableNode carries an IsInteger flag taken from
it and the declaration follows. That alone turned the float loop counters into
int ones, in six shaders.

and, or and xor are polymorphic - bitwise on integers, logical on comparison
results - and getting it wrong is silent. Listing them as integer opcodes made
a logical and read its float constant as an integer, printing 1.0 as
1065353216. GetSourceCount doubles as the is-an-integer-instruction test in
FindIntegerRegisters, so giving them an operand count there seeded the set
anyway; those roles are now separate. Deciding by the registers they touch
needs a fixed point, since xor feeding an or whose result reaches utof takes
two rounds, and it has to be per component: propagating per instruction typed
the float accumulator of mov r0.xy, l(0,0,0,0) as an integer because the
counter beside it was one. The AST already decided by shape - neither operand
a condition means bitwise - so the old refusal site was exactly that case.

Interpolation modifiers were read from the operand token instead of the opcode
token, so every pixel shader input came back linear and the modifier never
reached the HLSL. A nointerpolation input recompiled as a linear interpolant
is a different shader. SV_Position in a pixel shader is linear noperspective,
which fxc confirms, so absolute_multiply was recording the broken decode.

ps_2_0 takes its texture coordinates from t# registers declared as `dcl t0.xy`.
The semantic is the register number, the assembly writes dcl bare, and the
register is a method input rather than a temp. RegisterType.Addr and
RegisterType.Texture are one number, so a pixel shader reading t0 looked like
a vertex shader writing the address register - which is why it printed a0 and
declared a stray int a0.

Structured buffers wider than one component now work. StoreStructuredStatement
carries one value per component, read from the destination write mask rather
than a hardcoded component index of 3, and neither load nor store puts a
swizzle on the buffer name. A statement also holds nodes outside its input and
output maps, which the phi lowering could not see, so a vector element read
across a loop reached compilation unlowered; ReplaceInStatementNodes covers
the store and the clip alike.

Added ps_2_0/tex2d, ps_2_0/lerp, ps_4_0/interpolation, ps_4_0/clip_loop,
cs_4_1/compute_vector, cs_4_1/compute_loop and vs_4_0/bitwise.

+++ HlslDecompiler.Tests/ShaderSources/gs_4_1/circle.fx
@@ -18,7 +18,7 @@ void main(point GS_IN i[1], inout TriangleStream<GS_OUT> stream)
 	o.sv_position = i[0].sv_position;
 	o.color = i[0].color;
 	stream.Append(o);
-	for (float t0 = 1; t0 <= 17; t0 = t0 + 1) {
+	for (int t0 = 1; t0 <= 17; t0 = t0 + 1) {
 		o.sv_position = float4(0.5 * cos(0.392699093 * t0) + i[0].sv_position.x, 0.5 * sin(0.392699093 * t0) + i[0].sv_position.y, i[0].sv_position.zw);
 		o.color = i[0].color;
 		stream.Append(o);
