+++ HlslDecompiler.Tests/ShaderSources/gs_4_1/point_expand.fx
@@ -14,7 +14,7 @@ void main(point GS_IN i[1], inout TriangleStream<GS_OUT> stream)
 {
 	GS_OUT o;
 
-	for (float t0 = 0; t0 < 4; t0 = t0 + 1) {
+	for (int t0 = 0; t0 < 4; t0 = t0 + 1) {
 		o.sv_position = float4(t0 + i[0].sv_position.xy, i[0].sv_position.zw);
 		o.texcoord = t0;
 		stream.Append(o);
